Add time picker to select benchmark chart dates.
* Also move benchmark result components to a dedicated container.
