+++ client/package.json
@@ -8,7 +8,9 @@
     "@testing-library/user-event": "^7.2.1",
     "@types/highcharts": "^7.0.0",
     "@types/jest": "^24.9.1",
+    "@types/moment": "^2.13.0",
     "@types/node": "^12.12.46",
+    "@types/rc-time-picker": "^3.4.1",
     "@types/react": "^16.9.35",
     "@types/react-bootstrap": "^1.0.1",
     "@types/react-dom": "^16.9.8",
@@ -18,7 +20,9 @@
     "highcharts": "^8.1.1",
     "highcharts-react-official": "^3.0.0",
     "http-proxy-middleware": "^1.0.4",
+    "moment": "^2.27.0",
     "mutationobserver-shim": "^0.3.7",
+    "rc-time-picker": "^3.7.3",
     "react": "^16.13.1",
     "react-bootstrap": "^1.0.1",
     "react-day-picker": "^7.4.8",
