+++ client/src/App.tsx
@@ -1,5 +1,6 @@
 import 'bootstrap/dist/css/bootstrap.min.css';
 import 'react-day-picker/lib/style.css';
+import 'rc-time-picker/assets/index.css';
 import './App.css';
 
 import React from 'react';
