+++ client/src/Benchmark/BenchmarkResult/AssetsTable.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Table from 'react-bootstrap/Table';
 
-import { Asset } from '../types';
+import { Asset } from '../../types';
 
 interface Props {
   assets: Asset[]
