+++ client/src/Benchmark/BenchmarkResult/PricesAnalysisTable.tsx
