+++ client/src/Benchmark/BenchmarkResult/PricesStatisticsAnalysisTable.tsx
